Align simple example with go-backend rules
Split the runnable example into a thin main plus strategy and
memory broker files, wrap errors with action tags, and avoid
shadowing the memoryPosition type name.
